REFACTOR: dto위치 수정 및 인터페이스 이름 수정 See also: #11
