+++ src/main/java/com/sparta/deliveryi/ai/domain/service/AiLogRegister.java
@@ -3,5 +3,5 @@ package com.sparta.deliveryi.ai.domain.service;
 import com.sparta.deliveryi.ai.domain.AiLog;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AiLogRepository extends JpaRepository<AiLog, Long> {
+public interface AiLogRegister extends JpaRepository<AiLog, Long> {
 }
